ref(*): Refactors krustlet to have multiple crates
This also adds some of the base test files and dylibs for future refactor
to bring in the code from our CRI experiment
