+++ crates/kubelet/src/lib.rs
@@ -1,0 +1,12 @@
+#[macro_use]
+extern crate serde_json;
+
+#[macro_use]
+extern crate failure;
+
+pub mod kubelet;
+pub mod node;
+pub mod pod;
+mod server;
+
+pub use kubelet::*;
