+++ crates/wascc-provider/Cargo.toml
@@ -1,0 +1,26 @@
+[package]
+name = "wascc-provider"
+version = "0.1.0"
+authors = [
+    "Matt Butcher <matt.butcher@microsoft.com>",
+    "Matthew Fisher <matt.fisher@microsoft.com>",
+    "Radu Matei <radu.matei@microsoft.com>",
+    "Taylor Thomas <taylor.thomas@microsoft.com>",
+    "Brian Ketelsen <Brian.Ketelsen@microsoft.com>",
+    "Brian Hardock <Brian.Hardock@microsoft.com>",
+    "Ryan Levick <rylevick@microsoft.com>",
+]
+edition = "2018"
+
+[dependencies]
+failure = "0.1"
+wasmtime = "0.8"
+wasmtime-wasi = "0.8"
+wasi-common = "0.8"
+wascc-host = "0.3.2"
+log = "0.4"
+serde = "1.0"
+serde_derive = "1.0"
+serde_json = "1.0"
+kube = { version = "0.16.1", features = ["openapi"] }
+kubelet = { path = "../kubelet", version = "0.1.0" }
