+++ src/krustlet-wasi.rs
@@ -1,0 +1,3 @@
+fn main() -> Result<(), failure::Error> {
+    Ok(())
+}
